Bot/requirements.txt
python-telegram-bot==20.3
motor==3.3.2
python-dotenv==1.0.0
pymongo==4.6.0
watchdog==3.0.0
aiohttp==3.9.1
